shop db design
DAO pattern
