+++ app/src/main/java/com/intplus/shoppingspace/AllShopsActivity.java
@@ -1,12 +1,8 @@
 package com.intplus.shoppingspace;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.view.View;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
@@ -17,10 +13,8 @@ import android.view.Menu;
 import android.view.MenuItem;
 
 import com.intplus.shoppingspace.adapters.AllShopsListAdapter;
-import com.intplus.shoppingspace.adapters.DashboardGridAdapter;
 import com.intplus.shoppingspace.controller.AppController;
-import com.intplus.shoppingspace.controller.MainActivityController;
-import com.intplus.shoppingspace.helpers.Shop;
+import com.intplus.shoppingspace.model.Shop;
 
 import java.util.ArrayList;
