+++ app/src/main/java/com/intplus/shoppingspace/adapters/DashboardGridAdapter.java
@@ -7,11 +7,10 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import com.intplus.shoppingspace.R;
 import com.intplus.shoppingspace.app.AppConstants;
-import com.intplus.shoppingspace.helpers.Shop;
+import com.intplus.shoppingspace.model.Shop;
 
 import java.util.ArrayList;
